(hel-312) script checkReclamationsEtablissements cron
