+++ .gitignore
@@ -39,6 +39,7 @@ yarn-error.log*
 data_test
 data_set/.ssh
 data_set/sirec
+data_set/checked_sirec
 sshd_config
 ssh_key
