--- SIREC/sirec_202401231233_mini.csv
@@ -1,14 +1,0 @@
-"IDENTIFIANT";"NDEG_FINESS_RPPS";"ANNEE_DE_RECEPTION";"ENCOURS_NB_RECLA_TOTAL";"ENCOURS_NB_RECLA_MOTIF_10";"ENCOURS_NB_RECLA_MOTIF_11";"ENCOURS_NB_RECLA_MOTIF_12";"ENCOURS_NB_RECLA_MOTIF_13";"ENCOURS_NB_RECLA_MOTIF_14";"ENCOURS_NB_RECLA_MOTIF_15";"ENCOURS_NB_RECLA_MOTIF_16";"ENCOURS_NB_RECLA_MOTIF_17";"ENCOURS_NB_RECLA_MOTIF_18";"ENCOURS_NB_RECLA_MOTIF_19";"ENCOURS_NB_RECLA_MOTIF_155";"ENCOURS_NB_RECLA_MOTIF_156";"CLOT_NB_RECLA_TOTAL";"CLOT_NB_RECLA_MOTIF_10";"CLOT_NB_RECLA_MOTIF_11";"CLOT_NB_RECLA_MOTIF_12";"CLOT_NB_RECLA_MOTIF_13";"CLOT_NB_RECLA_MOTIF_14";"CLOT_NB_RECLA_MOTIF_15";"CLOT_NB_RECLA_MOTIF_16";"CLOT_NB_RECLA_MOTIF_17";"CLOT_NB_RECLA_MOTIF_18";"CLOT_NB_RECLA_MOTIF_19";"CLOT_NB_RECLA_MOTIF_155";"CLOT_NB_RECLA_MOTIF_156"
-"0100000652021";"010000065";"2021";;;;;;;;;;;;;;1;;;;;;;;;;;;1
-"0100004952021";"010000495";"2021";;;;;;;;;;;;;;2;1;;;;;1;;;;;;
-"0100017412023";"010001741";"2023";1;;;;1;;;;;;;;;;;;;;;;;;;;;
-"0100022282021";"010002228";"2021";;;;;;;;;;;;;;1;;;;;;;;;;;;1
-"0100032182023";"010003218";"2023";1;;;1;;;;;;;;;;;;;;;;;;;;;;
-"0100049012022";"010004901";"2022";;;;;;;;;;;;;;1;;;;;;1;;;;;;
-"0100052392022";"010005239";"2022";;;;;;;;;;;;;;5;;;2;;;3;;;;;;
-"0100053792021";"010005379";"2021";;;;;;;;;;;;;;1;;;1;;;;;;;;;
-"0100053792022";"010005379";"2022";;;;;;;;;;;;;;1;;;1;;;;;;;;;
-"0100054522022";"010005452";"2022";;;;;;;;;;;;;;1;;;;;;;;;;;;1
-"0100074662023";"010007466";"2023";1;;;;1;;;;;;;;;;;;;;;;;;;;;
-"0100081752023";"010008175";"2023";1;;1;;;;;;;;;;;;;;;;;;;;;;;
-"1111111111111";"010008175";"2023";1;;1;;;;;;;;;;;;;;;;;;;;;;;
