+++ download_data_source/métier/gateways/ControleDonnesSirecLoader.ts
@@ -1,0 +1,3 @@
+export interface ControleDonneesSirecLoader {
+    checkDowloadedSirecFile(): Promise<void>;
+}
