+++ package.json
@@ -57,6 +57,7 @@
     "updateCapacitéSanitaireEntitéJuridique:local": "pipenv run populate_capacites_sanitaire_entites_juridiques",
     "updateReclamationsEtablissements": "python -m datacrawler.import_reclamations",
     "updateReclamationsEtablissements:local": "pipenv run populate_reclamations",
+    "checkReclamationsEtablissements": "ts-node -T download_data_source/infrastructure/controllers/controleLesDonnesSirecCron.ts",
     "start": "next start",
     "start:db": "docker-compose down && docker-compose up -d && bash -c \"while ! docker-compose logs | grep -q 'listening on IPv4 address'; do echo 'Waiting for database'; sleep 0.5; done\"; yarn migrations:up",
     "test": "yarn test:typescript && yarn test:python",
@@ -82,6 +83,7 @@
     "@types/dompurify": "^3.0.5",
     "@types/joi": "^17.2.3",
     "@types/next-auth": "^3.15.0",
+    "@types/papaparse": "^5.3.14",
     "bcrypt": "^5.1.0",
     "chart.js": "^3.8.0",
     "chartjs-plugin-datalabels": "^2.0.0",
@@ -99,6 +101,7 @@
     "next-auth": "^4.22.1",
     "next-usequerystate": "^1.12.2",
     "node-fetch": "2.6.7",
+    "papaparse": "^5.4.1",
     "pg": "^8.7.3",
     "react": "18.0.0",
     "react-chartjs-2": "^4.2.0",
