--- src/backend/métier/use-cases/ReclamationUseCase.ts
@@ -1,9 +1,0 @@
-import { ReclamationLoader } from "../gateways/ReclamationLoader";
-
-export class ReclamationUseCase {
-  constructor(private reclamationLoader: ReclamationLoader) {}
-
-  async createReclamation(row: any): Promise<void> {
-    return this.reclamationLoader.createReclamation(row);
-  }
-}
